requirements.txt
flask==3.1.1
ultralytics==8.3.140
torch==2.7.0+cu128
opencv-python==4.9.0.80
pillow==10.4.0
numpy==1.26.4